move `TestContext` to `test_util.rs`.
