+++ df_engine_extensions/src/dist_sql_query/mod.rs
@@ -27,6 +27,8 @@ use table_engine::{
 pub mod codec;
 pub mod physical_plan;
 pub mod resolver;
+#[cfg(test)]
+pub mod test_util;
 
 /// Remote datafusion physical plan executor
 #[async_trait]
